% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organism.feature.R
\name{organism.feature}
\alias{organism.feature}
\alias{redscale.aphytis}
\alias{redscale.comperiella}
\alias{redscale.encarsia}
\title{Host-Parasitoid Interaction Table}
\format{
Columns of \code{organism.feature} are:
\describe{
  \item{ovip}{risk of life stage being selected for oviposition (egg-laying)}
  \item{feed}{risk of life stage being selected for feeding}
  \item{offspring}{Poisson mean number of offspring load for new parasite reared in host of this stage}
  \item{male}{risk of parasite laying a male egg in this host stage}
}
}
\description{
Host-Parasitoid Interaction Table for by Quantitative Population Ethology
simulation of Red Scale-Aphytis-Comperiella-Encarsia System
}
\details{
Parasites select hosts for ovipositing (egg-laying) or feeding based on
their stage of development. The risk of male offspring depends on stage, as
does the initial offspring load for the successfully reared parasite. These
tables are internalized via \code{init.simulation}.
}
\examples{

data(organism.feature)

}
\references{
\url{https://www.stat.wisc.edu/~yandell/ewing}
}
\seealso{
\code{\link[ewing]{init.simulation}}
}
\keyword{{datasets}}
