% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemperatureBase.R
\name{TemperatureBase}
\alias{TemperatureBase}
\alias{Temperature}
\alias{TemperaturePar}
\title{Temperature Base and Par}
\format{
\code{TemperatureBase} and \code{TemperaturePar} is a dataset described below:
\describe{The columns of \code{TemperaturePar} are:
  \item{Day}{day on which baseline begins}
  \item{Time}{time in hours from midnight}
  \item{Base}{temperature offset in percent between daily lows and highs}
}
\describe{The rows of \code{TemperaturePar} are:
 \item{Unit}{hours in day}
 \item{Days}{number of days}
 \item{Min}{minimum temperature for degree-day calculation (Fahrenheit)}
 \item{LowBeg}{begining daily low temperature}
 \item{LowEnd}{ending daily low temperature}
 \item{HighBeg}{begining daily high temperature}
 \item{HighEnd}{ending daily high temperature}
 \item{Length}{number of nodes for interpolating spline}
}
}
\description{
Data tables and features used by Quantitative Population Ethology simulation
of Red Scale-Aphytis-Comperiella-Encarsia System
}
\details{
Used internally by \code{init.simulation} to initialize temperature regime
for simulation.
}
\examples{

data(TemperatureBase)
data(TemperaturePar)

}
\references{
\url{https://www.stat.wisc.edu/~yandell/ewing}
}
\seealso{
\code{\link[ewing]{init.simulation}}
}
\keyword{{datasets}}
